requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-cloud-pubsub==2.19.0
google-cloud-firestore==2.14.0
gunicorn==21.2.0
base58
solders
